[DirectionProvider] support undefined direction
When rendering user-generated content, the direction could sometimes be
undefined. For example, a phone number or an empty string would not be strong
RTL or strong LTR text.

Before this PR, developers would have to check the direction and conditionally
render a DirectionProvider, which can be combersome.

With this PR `direction` can be undefined, in effect making DirectionProvider a
no-op. Decendent components that use `withDirection` would "skip" undefined
directions and use the context from the closest ancestor that has a defined
direction context (to be verified).