[DirectionProvider] add support for inherited direction
When rendering user-generated content, the direction could sometimes be
undefined. For example, a phone number or an empty string would not be
strong RTL/LTR text.

Before this PR, developers would have to check the direction and
conditionally render a DirectionProvider, which can be cumbersome.

This updates `withDirection` to take an optional `direction` prop to
override the `direction` from context, and wraps `DirectionProvider` in
the HOC. When `direction` is null, `DirectionProvider` gets the value
from context. In all other cases, the behavior is unchanged.